Rename the file to better reflect its contents
